Revert "Telephony: Convert subscription ID to int from long"
This reverts commit 577ba7cf9371e5fa1c7283c6aa118acfa25332d6.

+++ telecomm/java/android/telecom/TelecomManager.java
@@ -929,10 +929,10 @@ public class TelecomManager {
      * Returns current active subscription.
      * @hide
      */
-    public int getActiveSubscription() {
+    public long getActiveSubscription() {
         try {
             if (isServiceConnected()) {
-                return (int)getTelecomService().getActiveSubscription();
+                return getTelecomService().getActiveSubscription();
             }
         } catch (RemoteException e) {
             Log.e(TAG, "RemoteException attempting to get the active subsription.", e);
@@ -944,7 +944,7 @@ public class TelecomManager {
      * switches to other active subscription.
      * @hide
      */
-    public void switchToOtherActiveSub(int subId) {
+    public void switchToOtherActiveSub(long subId) {
         try {
             if (isServiceConnected()) {
                 getTelecomService().switchToOtherActiveSub(subId);
